Updated API base URL and fixed analyze endpoint

+++ tasks/models.py
@@ -3,8 +3,10 @@ from django.db import models
 class Task(models.Model):
     title = models.CharField(max_length=200)
     due_date = models.DateField()
-    importance = models.IntegerField(default=5)  # 1–10 scale
-    estimated_hours = models.IntegerField(default=1)
+    importance = models.IntegerField(default=5)  # 1–10
+    estimated_hours = models.FloatField(default=1)
+
+    # Store dependency IDs or names as a list
     dependencies = models.JSONField(default=list, blank=True)
 
     def __str__(self):
